debugger/renderer: Cast scaled size from float to uint32_t
Bye warning!

+++ src/emu/debugger/src/renderer/renderer.cpp
@@ -73,8 +73,8 @@ namespace eka2l1 {
         const std::uint32_t width, const std::uint32_t height, const std::uint32_t fb_width,
         const std::uint32_t fb_height) {
 
-        const std::uint32_t scaled_width = width / debugger_->get_config()->ui_scale,
-            scaled_height = height / debugger_->get_config()->ui_scale;
+        const std::uint32_t scaled_width = static_cast<std::uint32_t>(width / debugger_->get_config()->ui_scale),
+            scaled_height = static_cast<std::uint32_t>(height / debugger_->get_config()->ui_scale);
 
         if (!background_change_path_.empty()) {
             // Update the background
